requirements.txt
annotated-types==0.7.0
anyio==4.12.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
fastapi==0.109.0
filelock==3.20.1
filetype==1.2.0
fsspec==2025.12.0
google-ai-generativelanguage==0.4.0
google-api-core==2.28.1
google-auth==2.45.0
google-genai==1.56.0
google-generativeai==0.3.2
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.5
langchain-google-genai==4.1.2
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.1
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.0
opencv-python==4.8.1.78
orjson==3.11.5
ormsgpack==1.12.1
packaging==23.2
Pillow==10.1.0
pinecone-client==3.0.0
platformdirs==4.5.1
pluggy==1.6.0
proto-plus==1.27.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
pytest==7.4.4
pytest-asyncio==0.23.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.7.0
scenedetect==0.6.2
sentencepiece==0.2.0
sniffio==1.3.1
starlette==0.35.1
sympy==1.14.0
tenacity==8.5.0
tokenizers==0.19.1
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
transformers==4.44.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid6==2024.7.10
uuid_utils==0.12.0
uvicorn==0.27.0
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
zstandard==0.25.0
